requirements.txt
openai-agents==0.4.2
streamlit==1.51.0
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
qdrant-client==1.15.1
fastembed==0.7.3
langchain==1.0.3
langchain-community==0.4.1
langchain-openai==1.0.2
openai==2.7.1
langchain-text-splitters==1.0.0
python-dotenv==1.0.0
pypdf==6.6.0